fixing some bugs

+++ delete_db.sh
@@ -2,7 +2,10 @@
 
 function delete_db(){
 	read -p "Please Enter Database Name: " DBName
-	
+	if ! validate_database_name "$DBName"; then
+        return 
+    fi
+
 	if [ -d "Databases/$DBName" ]
 	then
 		rm -r "Databases/$DBName"
